[rocRoller] Preload and Sort Kernel Arguments (#4636)
## Motivation

AIROCROLL-1226 and AIROCROLL-1229

Speeds up small kernels by:

- Preloading kernel arguments where possible
- Ordering kernel arguments by first use so that we can delay the
`s_waitcnt` instruction as long as possible in the kernel
- Using wide `s_load` instructions to load multiple kernel arguments at
once.

## Technical Details

- Added new `MaxPreloadedKernargs` capability to enable this feature
where supported.
- We now use the terms eager and lazy to refer to when we are manually
loading kernel arguments as preloading has a different meaning.
- Deleted the launch-time only argument mechanism as it was already no
longer needed.

## Test Plan

 - Added tests for new and modified functionality.
     - Added new test that comprehensively tests argument loading
 - Added new rrsuite entries for small kernels.

+++ shared/rocroller/lib/source/Costs/LinearWeightedCost.cpp
@@ -328,11 +328,13 @@ namespace rocRoller
                 lgkmcnt    = 1 + maxLgkmcnt - cnt;
             }
 
-            int vmQueueLen = status.waitLengths.at(GPUWaitQueueType::LoadQueue)
-                             + status.waitLengths.at(GPUWaitQueueType::StoreQueue);
+            int vmQueueLen
+                = status.waitLengths.at(static_cast<size_t>(GPUWaitQueueType::LoadQueue))
+                  + status.waitLengths.at(static_cast<size_t>(GPUWaitQueueType::StoreQueue));
             float vectorQueueSat = std::max(vmQueueLen - m_weights.vmQueueLen, 0);
-            int   kmQueueLen     = status.waitLengths.at(GPUWaitQueueType::SMemQueue)
-                             + status.waitLengths.at(GPUWaitQueueType::DSQueue);
+            int   kmQueueLen
+                = status.waitLengths.at(static_cast<size_t>(GPUWaitQueueType::SMemQueue))
+                  + status.waitLengths.at(static_cast<size_t>(GPUWaitQueueType::DSQueue));
             float ldsQueueSat = std::max(kmQueueLen - m_weights.lgkmQueueLen, 0);
 
             float newSGPRs = status.allocatedRegisters.at(Register::Type::Scalar);
